Did some code refactoring and cleanup.
